ecommerce_classifier/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.7
decorator==5.1.1
dj-database-url==2.0.0
Django==4.1.7
django-cors-headers==4.0.0
django-environ==0.9.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
executing==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.23.1
ipython==8.13.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wcwidth==0.2.6
